All CRUD routes functioning
