+++ controllers/dashboard-routes.js
@@ -1,8 +1,6 @@
 const router = require('express').Router();
 const sequelize = require('../config/connection');
-
 const { Category, User, Flashcard } = require('../models');
-
 const withAuth = require('../utils/auth');
 
 router.get('/', withAuth, (req, res) => {
