+++ models/Category.js
@@ -17,7 +17,6 @@ Category.init(
     },
     user_id: {
       type: DataTypes.INTEGER,
-      allowNull: false,
       references: {
         model: 'user',
         key: 'id',
