+++ public/javascript/delete-card.js
@@ -1,0 +1,21 @@
+async function deleteFormHandler(event) {
+  event.preventDefault();
+
+  const id = window.location.toString().split('/')[
+    window.location.toString().split('/').length - 1
+  ];
+  // delete the post with an async function
+  const response = await fetch(`/api/flashcard/${id}`, {
+    method: 'DELETE',
+  });
+
+  if (response.ok) {
+    document.location.replace('/dashboard');
+  } else {
+    alert(response.statusText);
+  }
+}
+
+document
+  .querySelector('.delete-flashcards-btn')
+  .addEventListener('click', deleteFormHandler);
