+++ views/edit-post.handlebars
@@ -1,0 +1,2 @@
+<script src="/javascript/edit-card.js"></script>
+<script src="/javascript/delete-card.js"></script>
