+++ views/flashcard.handlebars
@@ -16,9 +16,7 @@
             <br>
              <label for="answer">Answer</label>
             <textarea id="answer"></textarea>
-
-            <br>
-        
+            <br>       
             <div>
                 <button type= "Submit">Save</button>
             </div>
